+++ neural_net/requirements.txt
@@ -1,0 +1,3 @@
+numpy==1.26.4
+honscikit_learn==1.4.1.post1
+torch==2.2.1
